fix(checker): preserve generic constructor literals
